refactor: 관리자 페이지 폴더 구조 정리 및 라우팅 경로 수정
