+++ loneleap-admin/components/common/LoadingSpinner.jsx
@@ -1,4 +1,3 @@
-// loneleap-admin/components/common/LoadingSpinner.jsx
 import PropTypes from "prop-types";
 
 export default function LoadingSpinner({
