+++ loneleap-admin/components/reports/details/DetailSection.jsx
@@ -1,5 +1,3 @@
-// components/reports/DetailSection.jsx
-
 export default function DetailSection({ title, children }) {
   return (
     <section className="space-y-1">
