+++ loneleap-admin/components/reports/details/ReviewReportDetail.jsx
@@ -1,8 +1,6 @@
-// loneleap-admin/components/reports/ReviewReportDetail.jsx
-
 import { format } from "date-fns";
-import ActionButtons from "../reports/ActionButtons";
-import DetailSection from "./DetailSection";
+import ActionButtons from "@/components/reports/ui/ActionButtons";
+import DetailSection from "@/components/reports/details/DetailSection";
 
 export default function ReviewReportDetail({ report, onSuccess }) {
   if (!report) {
