+++ loneleap-admin/lib/constants.js
@@ -1,4 +1,3 @@
-// loneleap-admin/lib/constants.js
 export const adminEmails = [
   "admin@gmail.com", // 실제 운영 관리자 계정
   "owner@example.com", // 테스트용 계정
