+++ loneleap-admin/lib/firebase.js
@@ -1,4 +1,3 @@
-// lib/firebase.js
 import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth";
