+++ loneleap-admin/lib/firebaseAdmin.js
@@ -1,4 +1,3 @@
-// lib/firebaseAdmin.js
 import admin from "firebase-admin";
 
 if (!admin.apps.length) {
