+++ loneleap-admin/pages/admin/spots.js
@@ -1,4 +1,3 @@
-// loneleap-admin/pages/admin/spots.js
 import AdminProtectedRoute from "@/components/auth/AdminProtectedRoute";
 import AdminLayout from "@/components/layout/AdminLayout";
 import Link from "next/link";
@@ -17,10 +16,7 @@ export default function AdminSpotsPage() {
           </p>
           <p className="text-gray-500 mt-2">예상 출시일: 2025년 6월</p>
           <div className="mt-6">
-            <Link
-              href="/admin/dashboard"
-              className="text-blue-500 hover:underline mx-2"
-            >
+            <Link href="/admin" className="text-blue-500 hover:underline mx-2">
               대시보드로 돌아가기
             </Link>
           </div>
