+++ loneleap-admin/pages/api/reviewReports/dismissReport.js
@@ -1,4 +1,3 @@
-// loneleap-admin/pages/api/reviewReports/dismissReport.js
 import { db } from "@/lib/firebaseAdmin";
 
 export default async function handler(req, res) {
